выносит типы пропсов в отдельный файл, единообразная деструктуризация пропсов, выносит константы в отдельный файл
