+++ src/common-prop-types.js
@@ -1,0 +1,12 @@
+import PropTypes from "prop-types";
+import {RoomType} from "./constants";
+
+export const CardType = PropTypes.exact({
+  image: PropTypes.string.isRequired,
+  price: PropTypes.number.isRequired,
+  rating: PropTypes.number.isRequired,
+  name: PropTypes.string.isRequired,
+  type: PropTypes.oneOf([RoomType.apartment, RoomType.hotel, RoomType.house, RoomType.room]).isRequired,
+  isPremium: PropTypes.bool.isRequired,
+  isFavorite: PropTypes.bool.isRequired,
+});
