+++ src/constants.js
@@ -1,0 +1,6 @@
+export const RoomType = {
+  apartment: `Apartment`,
+  room: `Private room`,
+  house: `House`,
+  hotel: `Hotel`,
+};
